requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.4.0
argcomplete==3.1.2
asgiref==3.8.1
attrs==23.2.0
base58==2.1.1
bitarray==3.0.0
bleak==0.22.3
blinker==1.9.0
boto3==1.34.131
botocore==1.34.138
cachetools==5.3.3
certifi==2024.6.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
construct==2.10.70
construct-classes==0.1.2
contourpy==1.3.1
cycler==0.12.1
cytoolz==1.0.0
distlib==0.3.8
distro==1.9.0
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.1
django-jsonfield==1.4.1
django-storages==1.14.3
djangorestframework==3.15.2
docutils==0.16
drf-yasg==1.21.7
ecdsa==0.19.0
ECPy==1.2.5
environ==1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==0.3.0
eth-typing==5.0.1
eth-utils==2.3.2
eth_abi==5.1.0
filelock==3.15.4
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
grpcio==1.68.0
grpcio-health-checking==1.59.0
grpcio-tools==1.59.0
gunicorn==23.0.0
h11==0.14.0
hexbytes==0.3.1
hidapi==0.14.0.post4
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
identify==2.6.0
idna==3.7
inflection==0.5.1
ipfshttpclient==0.4.13.2
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonrpcclient==4.0.3
jsonschema==4.0.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
ledgerblue==0.1.48
libusb1==3.1.0
lighthouseweb3==0.1.4
lru-dict==1.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
mnemonic==0.20
multiaddr==0.0.9
multidict==6.1.0
narwhals==1.14.2
ndeflib==0.3.3
netaddr==1.3.0
nfcpy==1.0.4
nodeenv==1.9.1
numpy==2.1.3
oauthlib==3.2.2
openai==1.35.10
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parsimonious==0.10.0
pillow==10.4.0
platformdirs==4.3.2
pre-commit==3.8.0
propcache==0.2.0
proto-plus==1.24.0
protobuf==4.21.6
psycopg2-binary==2.9.9
pyarrow==18.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycoin==0.92.20230326
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.8.0
pydantic_core==2.20.0
pydeck==0.9.1
pyDes==2.0.1
pyelftools==0.31
Pygments==2.18.0
pymultihash==0.8.2
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreBluetooth==10.3.1
pyobjc-framework-libdispatch==10.3.1
pyparsing==3.1.2
pyrsistent==0.20.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-gnupg==0.5.3
python-u2flib-host==3.0.3
pytz==2024.1
pyunormalize==16.0.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==2.0.0
rich==13.9.4
rlp==3.0.0
rpds-py==0.18.1
rsa==4.7.2
s3transfer==0.10.1
seaborn==0.13.2
selenium==4.22.0
setuptools==70.1.0
six==1.16.0
smmap==5.0.1
snet.cli==2.3.0
snet.contracts==0.1.1
snet.sdk==3.6.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
streamlit==1.40.1
tenacity==9.0.0
tokenizers==0.19.1
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.66.4
trezor==0.13.8
trio==0.26.0
trio-websocket==0.11.1
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.19
varint==1.0.2
virtualenv==20.26.4
watchdog==6.0.0
web3==6.11.1
websocket-client==1.8.0
websockets==14.1
wheel==0.41.2
whitenoise==6.7.0
wsproto==1.2.0
yarl==1.18.0
youtube-transcript-api==0.6.2
zipp==3.21.0
